Move column mapping to separate column mapper class
