+++ src/common/multi_file_reader.cpp
@@ -7,16 +7,11 @@
 #include "duckdb/function/function_set.hpp"
 #include "duckdb/function/table_function.hpp"
 #include "duckdb/main/config.hpp"
-#include "duckdb/planner/expression/bound_columnref_expression.hpp"
-#include "duckdb/planner/expression/bound_constant_expression.hpp"
+#include "duckdb/common/multi_file_reader_column_mapper.hpp"
 #include "duckdb/planner/expression/bound_reference_expression.hpp"
 #include "duckdb/planner/expression/bound_cast_expression.hpp"
-#include "duckdb/parser/expression/constant_expression.hpp"
 #include "duckdb/common/string_util.hpp"
-#include "duckdb/planner/filter/list.hpp"
-#include "duckdb/optimizer/statistics_propagator.hpp"
 #include <algorithm>
-#include <duckdb/parser/expression/cast_expression.hpp>
 
 namespace duckdb {
 
@@ -334,650 +329,15 @@ MultiFileReader::InitializeGlobalState(ClientContext &context, const MultiFileRe
 	return nullptr;
 }
 
-struct MultiFileIndexMapping {
-public:
-	explicit MultiFileIndexMapping(idx_t index) : index(index) {
-	}
-
-public:
-	MultiFileIndexMapping &AddMapping(idx_t from, idx_t to) {
-		auto res = child_mapping.emplace(from, make_uniq<MultiFileIndexMapping>(to));
-		return *res.first->second;
-	}
-
-public:
-	idx_t index;
-	unordered_map<idx_t, unique_ptr<MultiFileIndexMapping>> child_mapping;
-};
-
-struct MultiFileColumnMap {
-	MultiFileColumnMap(idx_t index, const LogicalType &local_type, const LogicalType &global_type)
-	    : mapping(index), local_type(local_type), global_type(global_type) {
-	}
-
-	MultiFileIndexMapping mapping;
-	const LogicalType &local_type;
-	const LogicalType &global_type;
-};
-
-struct ResultColumnMapping {
-	unordered_map<idx_t, MultiFileColumnMap> global_to_local;
-	string error;
-
-public:
-	bool HasError() const {
-		return !error.empty();
-	}
-};
-
-void PushColumnMapping(ClientContext &context, MultiFileFileReaderData &reader_data, BaseFileReader &reader,
-                       const LogicalType &global_type, const LogicalType &local_type, MultiFileLocalColumnId local_id,
-                       ResultColumnMapping &result, MultiFileGlobalIndex global_idx, const ColumnIndex &global_id) {
-	auto local_idx = reader.column_ids.size();
-
-	unique_ptr<Expression> expr;
-
-	ColumnIndex local_index(local_id.GetId());
-	if (reader.UseCastMap()) {
-		// reader is responsible for casting
-		expr = make_uniq<BoundReferenceExpression>(global_type, local_idx);
-		if (global_type != local_type) {
-			reader.cast_map[local_id.GetId()] = global_type;
-		}
-	} else {
-		expr = make_uniq<BoundReferenceExpression>(local_type, local_idx);
-		if (global_type != local_type) {
-			expr = BoundCastExpression::AddCastToType(context, std::move(expr), global_type);
-		} else {
-			//! FIXME: local fields are not guaranteed to match with the global fields for this struct
-			local_index = ColumnIndex(local_id.GetId(), global_id.GetChildIndexes());
-		}
-	}
-	reader_data.expressions.push_back(std::move(expr));
-
-	MultiFileColumnMap index_mapping(local_idx, local_type, global_type);
-	result.global_to_local.insert(make_pair(global_idx.GetIndex(), std::move(index_mapping)));
-	reader.column_ids.push_back(local_id);
-	reader.column_indexes.push_back(std::move(local_index));
-}
-
-ResultColumnMapping MultiFileReader::CreateColumnMappingByName(
-    ClientContext &context, MultiFileFileReaderData &reader_data,
-    const vector<MultiFileReaderColumnDefinition> &global_columns, const vector<ColumnIndex> &global_column_ids,
-    const MultiFileReaderBindData &bind_data, const virtual_column_map_t &virtual_columns) {
-	auto &reader = *reader_data.reader;
-	auto &local_columns = reader.GetColumns();
-	auto &file_name = reader.GetFileName();
-	// we have expected types: create a map of name -> (local) column id
-	case_insensitive_map_t<MultiFileLocalColumnId> name_map;
-	for (idx_t col_idx = 0; col_idx < local_columns.size(); col_idx++) {
-		auto &column = local_columns[col_idx];
-		name_map.emplace(column.name, MultiFileLocalColumnId(col_idx));
-	}
-	ResultColumnMapping result;
-	auto &expressions = reader_data.expressions;
-	// FIXME - this should be removed eventually
-	for (idx_t i = 0; i < global_column_ids.size(); i++) {
-		auto global_idx = MultiFileGlobalIndex(i);
-		// check if this is a constant column
-		optional_idx constant_idx;
-		for (idx_t j = 0; j < reader_data.constant_map.size(); j++) {
-			auto constant_index = MultiFileConstantMapIndex(j);
-			auto &entry = reader_data.constant_map[constant_index];
-			if (entry.column_idx.GetIndex() == i) {
-				constant_idx = j;
-				break;
-			}
-		}
-		if (constant_idx.IsValid()) {
-			// this column is constant for this file
-			auto constant_index = MultiFileConstantMapIndex(constant_idx.GetIndex());
-			auto &constant_entry = reader_data.constant_map[constant_index];
-			expressions.push_back(make_uniq<BoundConstantExpression>(constant_entry.value));
-			continue;
-		}
-		// not constant - look up the column in the name map
-		auto &global_id = global_column_ids[i];
-		auto global_column_id = global_id.GetPrimaryIndex();
-		if (IsVirtualColumn(global_column_id)) {
-			// virtual column - these are emitted for every file
-			auto virtual_entry = virtual_columns.find(global_column_id);
-			if (virtual_entry == virtual_columns.end()) {
-				throw InternalException("Virtual column id %d not found in virtual columns map", global_column_id);
-			}
-			expressions.push_back(make_uniq<BoundConstantExpression>(Value(virtual_entry->second.type)));
-			continue;
-		}
-		if (global_column_id >= global_columns.size()) {
-			throw InternalException(
-			    "MultiFileReader::CreateColumnMappingByName - global_id is out of range in global_types for this file");
-		}
-		auto &global_column = global_columns[global_column_id];
-		auto identifier = global_column.GetIdentifierName();
-		auto entry = name_map.find(identifier);
-		if (entry == name_map.end()) {
-			// identiier not present in file, use default value
-			if (global_column.default_expression) {
-				reader_data.constant_map.Add(global_idx, global_column.GetDefaultValue());
-				expressions.push_back(make_uniq<BoundConstantExpression>(global_column.GetDefaultValue()));
-				continue;
-			} else {
-				string candidate_names;
-				for (auto &column : local_columns) {
-					if (!candidate_names.empty()) {
-						candidate_names += ", ";
-					}
-					candidate_names += column.name;
-				}
-				result.error = StringUtil::Format(
-				    "Failed to read file \"%s\": schema mismatch in glob: column \"%s\" was read from "
-				    "the original file \"${INITIAL_FILE}\", but could not be found in file \"%s\".\nCandidate names: "
-				    "%s\nIf you are trying to "
-				    "read files with different schemas, try setting union_by_name=True",
-				    file_name, identifier, file_name, candidate_names);
-				return result;
-			}
-		}
-		// we found the column in the local file - check if the types are the same
-		auto local_id = entry->second;
-		D_ASSERT(global_column_id < global_columns.size());
-		D_ASSERT(local_id.GetId() < local_columns.size());
-		auto &global_type = global_columns[global_column_id].type;
-		auto &local_type = local_columns[local_id.GetId()].type;
-
-		PushColumnMapping(context, reader_data, reader, global_type, local_type, local_id, result, global_idx,
-		                  global_id);
-	}
-	D_ASSERT(global_column_ids.size() == reader_data.expressions.size());
-	return result;
-}
-
-ResultColumnMapping MultiFileReader::CreateColumnMappingByFieldId(
-    ClientContext &context, MultiFileFileReaderData &reader_data,
-    const vector<MultiFileReaderColumnDefinition> &global_columns, const vector<ColumnIndex> &global_column_ids,
-    const MultiFileReaderBindData &bind_data, const virtual_column_map_t &virtual_columns) {
-#ifdef DEBUG
-	//! Make sure the global columns have field_ids to match on
-	for (auto &column : global_columns) {
-		D_ASSERT(!column.identifier.IsNull());
-		D_ASSERT(column.identifier.type().id() == LogicalTypeId::INTEGER);
-	}
-#endif
-
-	auto &reader = *reader_data.reader;
-	auto &local_columns = reader.GetColumns();
-
-	ResultColumnMapping result;
-	// we have expected types: create a map of field_id -> column index
-	unordered_map<int32_t, MultiFileLocalColumnId> field_id_map;
-	for (idx_t col_idx = 0; col_idx < local_columns.size(); col_idx++) {
-		auto &column = local_columns[col_idx];
-		if (column.identifier.IsNull()) {
-			// Extra columns at the end will not have a field_id
-			break;
-		}
-		auto field_id = column.GetIdentifierFieldId();
-		field_id_map.emplace(field_id, MultiFileLocalColumnId(col_idx));
-	}
-
-	// loop through the schema definition
-	auto &expressions = reader_data.expressions;
-	for (idx_t i = 0; i < global_column_ids.size(); i++) {
-		auto global_idx = MultiFileGlobalIndex(i);
-
-		optional_idx constant_idx;
-		for (idx_t j = 0; j < reader_data.constant_map.size(); j++) {
-			auto constant_index = MultiFileConstantMapIndex(j);
-			auto &entry = reader_data.constant_map[constant_index];
-			if (entry.column_idx.GetIndex() == i) {
-				constant_idx = j;
-				break;
-			}
-		}
-		if (constant_idx.IsValid()) {
-			// this column is constant for this file
-			auto constant_index = MultiFileConstantMapIndex(constant_idx.GetIndex());
-			auto &constant_entry = reader_data.constant_map[constant_index];
-			expressions.push_back(make_uniq<BoundConstantExpression>(constant_entry.value));
-			continue;
-		}
-
-		// Handle any generate columns that are not in the schema (currently only file_row_number)
-		auto &global_id = global_column_ids[i];
-		auto global_column_id = global_column_ids[i].GetPrimaryIndex();
-
-		if (IsVirtualColumn(global_column_id)) {
-			// virtual column - these are emitted for every file
-			auto virtual_entry = virtual_columns.find(global_column_id);
-			if (virtual_entry == virtual_columns.end()) {
-				throw InternalException("Virtual column id %d not found in virtual columns map", global_column_id);
-			}
-			expressions.push_back(make_uniq<BoundConstantExpression>(Value(virtual_entry->second.type)));
-			continue;
-		}
-
-		auto local_idx = MultiFileLocalIndex(reader.column_ids.size());
-		if (global_column_id >= global_columns.size()) {
-			if (bind_data.file_row_number_idx == global_column_id) {
-				// FIXME: this needs a more extensible solution
-				auto new_column_id = MultiFileLocalColumnId(field_id_map.size());
-				reader.column_ids.push_back(new_column_id);
-				reader.column_indexes.emplace_back(field_id_map.size());
-				//! FIXME: what to do here???
-				expressions.push_back(make_uniq<BoundReferenceExpression>(LogicalType::BIGINT, local_idx));
-			} else {
-				throw InternalException("Unexpected generated column");
-			}
-			continue;
-		}
-
-		const auto &global_column = global_columns[global_column_id];
-		D_ASSERT(!global_column.identifier.IsNull());
-		auto it = field_id_map.find(global_column.GetIdentifierFieldId());
-		if (it == field_id_map.end()) {
-			// field id not present in file, use default value
-			auto &default_val = global_column.default_expression;
-			D_ASSERT(default_val);
-			if (default_val->type != ExpressionType::VALUE_CONSTANT) {
-				throw NotImplementedException("Default expression that isn't constant is not supported yet");
-			}
-			auto &constant_expr = default_val->Cast<ConstantExpression>();
-			expressions.push_back(make_uniq<BoundConstantExpression>(constant_expr.value));
-			reader_data.constant_map.Add(global_idx, constant_expr.value);
-			continue;
-		}
-
-		const auto &local_id = it->second;
-		auto &local_column = local_columns[local_id.GetId()];
-		auto &global_type = global_column.type;
-		auto &local_type = local_column.type;
-
-		PushColumnMapping(context, reader_data, reader, global_type, local_type, local_id, result, global_idx,
-		                  global_id);
-	}
-	D_ASSERT(global_column_ids.size() == reader_data.expressions.size());
-	return result;
-}
-
-ResultColumnMapping MultiFileReader::CreateColumnMapping(ClientContext &context, MultiFileFileReaderData &reader_data,
-                                                         const vector<MultiFileReaderColumnDefinition> &global_columns,
-                                                         const vector<ColumnIndex> &global_column_ids,
-                                                         const MultiFileReaderBindData &bind_data,
-                                                         const virtual_column_map_t &virtual_columns) {
-	switch (bind_data.mapping) {
-	case MultiFileReaderColumnMappingMode::BY_NAME: {
-		return CreateColumnMappingByName(context, reader_data, global_columns, global_column_ids, bind_data,
-		                                 virtual_columns);
-	}
-	case MultiFileReaderColumnMappingMode::BY_FIELD_ID: {
-		return CreateColumnMappingByFieldId(context, reader_data, global_columns, global_column_ids, bind_data,
-		                                    virtual_columns);
-	}
-	default: {
-		throw InternalException("Unsupported MultiFileReaderColumnMappingMode type");
-	}
-	}
-}
-
-static bool EvaluateFilterAgainstConstant(TableFilter &filter, const Value &constant) {
-	const auto type = filter.filter_type;
-
-	switch (type) {
-	case TableFilterType::CONSTANT_COMPARISON: {
-		auto &constant_filter = filter.Cast<ConstantFilter>();
-		if (constant.IsNull()) {
-			return false;
-		}
-		return constant_filter.Compare(constant);
-	}
-	case TableFilterType::IS_NULL: {
-		return constant.IsNull();
-	}
-	case TableFilterType::IS_NOT_NULL: {
-		return !constant.IsNull();
-	}
-	case TableFilterType::IN_FILTER: {
-		auto &in_filter = filter.Cast<InFilter>();
-		for (auto &val : in_filter.values) {
-			if (!constant.IsNull() && val == constant) {
-				return true;
-			}
-		}
-		return false;
-	}
-	case TableFilterType::CONJUNCTION_OR: {
-		auto &or_filter = filter.Cast<ConjunctionOrFilter>();
-		for (auto &it : or_filter.child_filters) {
-			if (EvaluateFilterAgainstConstant(*it, constant)) {
-				return true;
-			}
-		}
-		return false;
-	}
-	case TableFilterType::CONJUNCTION_AND: {
-		auto &and_filter = filter.Cast<ConjunctionAndFilter>();
-		auto res = make_uniq<ConjunctionAndFilter>();
-		for (auto &it : and_filter.child_filters) {
-			if (!EvaluateFilterAgainstConstant(*it, constant)) {
-				return false;
-			}
-		}
-		return true;
-	}
-	case TableFilterType::STRUCT_EXTRACT: {
-		auto &struct_filter = filter.Cast<StructFilter>();
-		auto &child_filter = struct_filter.child_filter;
-
-		if (constant.type().id() != LogicalTypeId::STRUCT) {
-			throw InternalException(
-			    "Constant for this column is not of type struct, but used in a STRUCT_EXTRACT TableFilter");
-		}
-		auto &struct_fields = StructValue::GetChildren(constant);
-		auto field_index = struct_filter.child_idx;
-		if (field_index >= struct_fields.size()) {
-			throw InternalException("STRUCT_EXTRACT looks for child_idx %d, but constant only has %d children",
-			                        field_index, struct_fields.size());
-		}
-		auto &field_name = StructType::GetChildName(constant.type(), field_index);
-		if (!StringUtil::CIEquals(field_name, struct_filter.child_name)) {
-			throw InternalException("STRUCT_EXTRACT looks for a child with name '%s' at index %d, but constant has a "
-			                        "field with '%s' as the name for that index",
-			                        struct_filter.child_name, field_index, field_name);
-		}
-		auto &child_constant = struct_fields[field_index];
-		return EvaluateFilterAgainstConstant(*child_filter, child_constant);
-	}
-	case TableFilterType::OPTIONAL_FILTER: {
-		auto &optional_filter = filter.Cast<OptionalFilter>();
-		if (optional_filter.child_filter) {
-			return EvaluateFilterAgainstConstant(*optional_filter.child_filter, constant);
-		}
-		return true;
-	}
-	case TableFilterType::DYNAMIC_FILTER: {
-		auto &dynamic_filter = filter.Cast<DynamicFilter>();
-		if (!dynamic_filter.filter_data) {
-			//! No filter_data assigned (does this mean the DynamicFilter is broken??)
-			return true;
-		}
-		if (!dynamic_filter.filter_data->initialized) {
-			//! Not initialized
-			return true;
-		}
-		if (!dynamic_filter.filter_data->filter) {
-			//! No filter present
-			return true;
-		}
-		lock_guard<mutex> lock(dynamic_filter.filter_data->lock);
-		return EvaluateFilterAgainstConstant(*dynamic_filter.filter_data->filter, constant);
-	}
-	default:
-		throw NotImplementedException("Can't evaluate TableFilterType (%s) against a constant",
-		                              EnumUtil::ToString(type));
-	}
-}
-
-namespace {
-
-struct EvaluationResult {
-	//! Whether evaluation of any of the filters against the global constants failed.
-	bool can_skip_file = false;
-	//! The remaining filters that need to be converted to local filters.
-	map<idx_t, reference<TableFilter>> remaining_filters;
-};
-
-} // namespace
-
-static EvaluationResult EvaluateConstantFilters(MultiFileFileReaderData &reader_data,
-                                                optional_ptr<TableFilterSet> filters,
-                                                const vector<MultiFileReaderColumnDefinition> &global_columns,
-                                                const vector<ColumnIndex> &global_column_ids,
-                                                const virtual_column_map_t &virtual_columns,
-                                                ResultColumnMapping &mapping) {
-	EvaluationResult result;
-	if (!filters) {
-		return result;
-	}
-	auto &global_to_local = mapping.global_to_local;
-	for (auto &it : filters->filters) {
-		auto &global_index = it.first;
-		auto &global_filter = it.second;
-
-		auto local_it = global_to_local.find(it.first);
-		if (local_it != global_to_local.end()) {
-			//! File has this column, filter needs to be evaluated later
-			result.remaining_filters.emplace(global_index, *global_filter);
-			continue;
-		}
-
-		//! FIXME: this does not check for filters against struct fields that are not present in the file
-		auto global_column_id = global_column_ids[global_index].GetPrimaryIndex();
-		Value constant_value;
-		auto virtual_it = virtual_columns.find(global_column_ids[global_index].GetPrimaryIndex());
-		if (virtual_it != virtual_columns.end()) {
-			auto &virtual_column = virtual_it->second;
-			if (virtual_column.name == "filename") {
-				constant_value = Value(reader_data.reader->GetFileName());
-			} else {
-				throw InternalException("Unrecognized virtual column found: %s", virtual_column.name);
-			}
-		} else {
-			bool has_constant = false;
-			for (idx_t i = 0; i < reader_data.constant_map.size(); i++) {
-				auto &constant_map_entry = reader_data.constant_map[MultiFileConstantMapIndex(i)];
-				if (constant_map_entry.column_idx.GetIndex() == global_index) {
-					has_constant = true;
-					constant_value = constant_map_entry.value;
-					break;
-				}
-			}
-			if (!has_constant) {
-				auto &global_column = global_columns[global_column_id];
-				throw InternalException(
-				    "Column '%s' is not present in the file, but no constant_map entry exists for it!",
-				    global_column.name);
-			}
-		}
-
-		if (!EvaluateFilterAgainstConstant(*global_filter, constant_value)) {
-			result.can_skip_file = true;
-			return result;
-		}
-	}
-
-	return result;
-}
-
-static unique_ptr<TableFilter> TryCastTableFilter(const TableFilter &global_filter, MultiFileIndexMapping &mapping,
-                                                  const LogicalType &target_type) {
-	auto type = global_filter.filter_type;
-
-	switch (type) {
-	case TableFilterType::CONJUNCTION_OR: {
-		auto &or_filter = global_filter.Cast<ConjunctionOrFilter>();
-		auto res = make_uniq<ConjunctionOrFilter>();
-		for (auto &it : or_filter.child_filters) {
-			auto child_filter = TryCastTableFilter(*it, mapping, target_type);
-			if (!child_filter) {
-				return nullptr;
-			}
-			res->child_filters.push_back(std::move(child_filter));
-		}
-		return std::move(res);
-	}
-	case TableFilterType::CONJUNCTION_AND: {
-		auto &and_filter = global_filter.Cast<ConjunctionAndFilter>();
-		auto res = make_uniq<ConjunctionAndFilter>();
-		for (auto &it : and_filter.child_filters) {
-			auto child_filter = TryCastTableFilter(*it, mapping, target_type);
-			if (!child_filter) {
-				return nullptr;
-			}
-			res->child_filters.push_back(std::move(child_filter));
-		}
-		return std::move(res);
-	}
-	case TableFilterType::STRUCT_EXTRACT: {
-		auto &struct_filter = global_filter.Cast<StructFilter>();
-		auto &child_filter = struct_filter.child_filter;
-
-		//! FXIME: The previous step should ensure that filters that target fields that are not present in the file are
-		//! evaluated earlier
-		//! For now we will assume the mapping is 1-to-1
-		MultiFileIndexMapping struct_mapping(struct_filter.child_idx);
-		auto &struct_type = StructType::GetChildTypes(target_type);
-		auto new_child_filter =
-		    TryCastTableFilter(*child_filter, struct_mapping, struct_type[struct_filter.child_idx].second);
-		if (!new_child_filter) {
-			return nullptr;
-		}
-		//! TODO: renaming fields should probably be respected here?
-		auto child_name = struct_filter.child_name;
-		return make_uniq<StructFilter>(struct_mapping.index, child_name, std::move(new_child_filter));
-	}
-	case TableFilterType::OPTIONAL_FILTER: {
-		auto &optional_filter = global_filter.Cast<OptionalFilter>();
-		auto child_result = TryCastTableFilter(*optional_filter.child_filter, mapping, target_type);
-		if (!child_result) {
-			return nullptr;
-		}
-		return make_uniq<OptionalFilter>(std::move(child_result));
-	}
-	case TableFilterType::DYNAMIC_FILTER: {
-		// we can't transfer dynamic filters over casts directly
-		// BUT we can copy the current state of the filter and push that
-		// FIXME: we could solve this in a different manner as well by pushing the dynamic filter directly
-		auto &dynamic_filter = global_filter.Cast<DynamicFilter>();
-		if (!dynamic_filter.filter_data) {
-			return nullptr;
-		}
-		if (!dynamic_filter.filter_data->initialized) {
-			return nullptr;
-		}
-		if (!dynamic_filter.filter_data->filter) {
-			return nullptr;
-		}
-		lock_guard<mutex> lock(dynamic_filter.filter_data->lock);
-		return TryCastTableFilter(*dynamic_filter.filter_data->filter, mapping, target_type);
-	}
-	case TableFilterType::IS_NULL:
-	case TableFilterType::IS_NOT_NULL:
-		// these filters can just be copied as they don't depend on type
-		return global_filter.Copy();
-	case TableFilterType::CONSTANT_COMPARISON: {
-		auto &constant_filter = global_filter.Cast<ConstantFilter>();
-		auto new_constant = constant_filter.constant;
-		if (!new_constant.DefaultTryCastAs(target_type)) {
-			return nullptr;
-		}
-		return make_uniq<ConstantFilter>(constant_filter.comparison_type, std::move(new_constant));
-	}
-	case TableFilterType::IN_FILTER: {
-		auto &in_filter = global_filter.Cast<InFilter>();
-		auto in_list = in_filter.values;
-		for (auto &val : in_list) {
-			if (!val.DefaultTryCastAs(target_type)) {
-				return nullptr;
-			}
-		}
-		return make_uniq<InFilter>(std::move(in_list));
-	}
-	default:
-		throw NotImplementedException("Can't convert TableFilterType (%s) from global to local indexes",
-		                              EnumUtil::ToString(type));
-	}
-}
-
-void SetIndexToZero(Expression &expr) {
-	if (expr.type == ExpressionType::BOUND_REF) {
-		auto &ref = expr.Cast<BoundReferenceExpression>();
-		ref.index = 0;
-		return;
-	}
-
-	ExpressionIterator::EnumerateChildren(expr, [&](Expression &child) { SetIndexToZero(child); });
-}
-
-unique_ptr<TableFilterSet> CreateFilters(ClientContext &context, MultiFileFileReaderData &reader_data,
-                                         const vector<MultiFileReaderColumnDefinition> &global_columns,
-                                         map<idx_t, reference<TableFilter>> &filters, ResultColumnMapping &mapping) {
-	if (filters.empty()) {
-		return nullptr;
-	}
-	auto &reader = *reader_data.reader;
-	auto &global_to_local = mapping.global_to_local;
-	auto result = make_uniq<TableFilterSet>();
-	for (auto &it : filters) {
-		auto &global_index = it.first;
-		auto &global_filter = it.second.get();
-
-		auto local_it = global_to_local.find(global_index);
-		if (local_it == global_to_local.end()) {
-			throw InternalException(
-			    "Error in 'EvaluateConstantFilters', this filter should not end up in CreateFilters!");
-		}
-		auto &map_entry = local_it->second;
-		auto local_id = map_entry.mapping.index;
-		auto filter_idx = reader.column_indexes[local_id].GetPrimaryIndex();
-		auto &local_type = map_entry.local_type;
-		auto &global_type = map_entry.global_type;
-
-		unique_ptr<TableFilter> local_filter;
-		if (local_type == global_type) {
-			// no conversion required - just copy the filter
-			local_filter = global_filter.Copy();
-		} else {
-			// types are different - try to convert
-			// first check if we can safely convert (i.e. if the conversion is lossless would not change the result)
-			if (StatisticsPropagator::CanPropagateCast(local_type, global_type)) {
-				// if we can convert - try to actually convert
-				local_filter = TryCastTableFilter(global_filter, map_entry.mapping, local_type);
-			}
-		}
-		if (local_filter) {
-			// succeeded in casting - push the local filter
-			result->filters.emplace(local_id, std::move(local_filter));
-		} else {
-			// failed to cast - copy the global filter and evaluate the conversion expression in the reader
-			result->filters.emplace(local_id, global_filter.Copy());
-
-			// add the expression to the expression map - we are now evaluating this inside the reader directly
-			// we need to set the index of the references inside the expression to 0
-			SetIndexToZero(*reader_data.expressions[local_id]);
-			reader.expression_map[filter_idx] = std::move(reader_data.expressions[local_id]);
-
-			// reset the expression - since we are evaluating it in the reader we can just reference it
-			reader_data.expressions[local_id] = make_uniq<BoundReferenceExpression>(global_type, local_id);
-		}
-	}
-	return result;
-}
-
 ReaderInitializeType MultiFileReader::CreateMapping(ClientContext &context, MultiFileFileReaderData &reader_data,
                                                     const vector<MultiFileReaderColumnDefinition> &global_columns,
                                                     const vector<ColumnIndex> &global_column_ids,
                                                     optional_ptr<TableFilterSet> filters, const string &initial_file,
                                                     const MultiFileReaderBindData &bind_data,
                                                     const virtual_column_map_t &virtual_columns) {
-	// copy global columns and inject any different defaults
-	auto result =
-	    CreateColumnMapping(context, reader_data, global_columns, global_column_ids, bind_data, virtual_columns);
-	if (result.HasError()) {
-		throw IOException(StringUtil::Replace(result.error, "${INITIAL_FILE}", initial_file));
-	}
-	//! Evaluate the filters against the column(s) that are constant for this file (not present in the local schema)
-	//! If any of these fail, the file can be skipped entirely
-	auto evaluation_result =
-	    EvaluateConstantFilters(reader_data, filters, global_columns, global_column_ids, virtual_columns, result);
-	if (evaluation_result.can_skip_file) {
-		return ReaderInitializeType::SKIP_READING_FILE;
-	}
-
-	reader_data.reader->filters =
-	    CreateFilters(context, reader_data, global_columns, evaluation_result.remaining_filters, result);
-
-	// for any remaining casts - push them as expressions
-	return ReaderInitializeType::INITIALIZED;
+	MultiFileReaderColumnMapper column_mapper(context, reader_data, global_columns, global_column_ids, filters,
+	                                          initial_file, bind_data, virtual_columns);
+	return column_mapper.CreateMapping();
 }
 
 string GetExtendedMultiFileError(const MultiFileBindData &bind_data, const Expression &expr, BaseFileReader &reader,
@@ -1002,38 +362,27 @@ string GetExtendedMultiFileError(const MultiFileBindData &bind_data, const Expre
 	string extended_error;
 	if (!bind_data.table_columns.empty()) {
 		// COPY .. FROM
-		extended_error = StringUtil::Format(
-		    "In file \"%s\" the column \"%s\" has type %s, but we are trying to load it into column ", reader.file_name,
-		    local_col.name, source_type);
+		string target_column;
 		if (expr_idx < bind_data.table_columns.size()) {
-			extended_error += "\"" + bind_data.table_columns[expr_idx] + "\" ";
+			target_column = "\"" + bind_data.table_columns[expr_idx] + "\" ";
 		}
-		extended_error += StringUtil::Format("with type %s.", target_type);
-		extended_error +=
-		    StringUtil::Format("\nThis means the %s schema does not match the schema of the table.", reader_type);
-		extended_error += "\nPossible solutions:";
-		extended_error += StringUtil::Format(
-		    "\n* Insert by name instead of by position using \"INSERT INTO tbl BY NAME SELECT * FROM "
-		    "%s(...)\"",
+		extended_error = StringUtil::Format(
+		    "In file \"%s\" the column \"%s\" has type %s, but we are trying to load it into column %swith type "
+		    "%s.\nThis means the %s schema does not match the schema of the table.\nPossible solutions:\n* Insert by "
+		    "name instead of by position using \"INSERT INTO tbl BY NAME SELECT * FROM %s(...)\"\n* Manually specify "
+		    "which columns to insert using \"INSERT INTO tbl SELECT ... FROM %s(...)\"",
+		    reader.file_name, local_col.name, source_type, target_column, target_type, reader_type, function_name,
 		    function_name);
-		extended_error +=
-		    StringUtil::Format("\n* Manually specify which columns to insert using \"INSERT INTO tbl SELECT ... FROM "
-		                       "%s(...)\"",
-		                       function_name);
 	} else {
 		// read_parquet() with multiple files
-		extended_error =
-		    StringUtil::Format("In file \"%s\" the column \"%s\" has type %s, but we are trying to read it as type %s.",
-		                       reader.file_name, local_col.name, source_type, target_type);
-		extended_error +=
-		    StringUtil::Format("\nThis can happen when reading multiple %s files. The schema information is taken from "
-		                       "the first %s file by default. Possible solutions:\n",
-		                       reader_type, reader_type);
-		extended_error +=
-		    StringUtil::Format("* Enable the union_by_name=True option to combine the schema of all %s files "
-		                       "(duckdb.org/docs/data/multiple_files/combining_schemas)\n",
-		                       reader_type);
-		extended_error += "* Use a COPY statement to automatically derive types from an existing table.";
+		extended_error = StringUtil::Format(
+		    "In file \"%s\" the column \"%s\" has type %s, but we are trying to read it as type %s."
+		    "\nThis can happen when reading multiple %s files. The schema information is taken from "
+		    "the first %s file by default. Possible solutions:\n"
+		    "* Enable the union_by_name=True option to combine the schema of all %s files "
+		    "(duckdb.org/docs/data/multiple_files/combining_schemas)\n",
+		    "* Use a COPY statement to automatically derive types from an existing table.", reader.file_name,
+		    local_col.name, source_type, target_type, reader_type, reader_type, reader_type);
 	}
 	first_message = StringUtil::Format("failed to cast column \"%s\" from type %s to %s: ", local_col.name, source_type,
 	                                   target_type);
