+++ src/include/duckdb/common/multi_file_reader.hpp
@@ -27,7 +27,6 @@ class LogicalGet;
 class Expression;
 class ClientContext;
 class DataChunk;
-struct ResultColumnMapping;
 
 enum class ReaderInitializeType { INITIALIZED, SKIP_READING_FILE };
 
@@ -228,22 +227,6 @@ public:
 	                                                       TableFunctionPartitionInput &input);
 
 protected:
-	virtual ResultColumnMapping CreateColumnMapping(ClientContext &context, MultiFileFileReaderData &reader_data,
-	                                                const vector<MultiFileReaderColumnDefinition> &global_columns,
-	                                                const vector<ColumnIndex> &global_column_ids,
-	                                                const MultiFileReaderBindData &bind_data,
-	                                                const virtual_column_map_t &virtual_columns);
-	virtual ResultColumnMapping
-	CreateColumnMappingByFieldId(ClientContext &context, MultiFileFileReaderData &reader_data,
-	                             const vector<MultiFileReaderColumnDefinition> &global_columns,
-	                             const vector<ColumnIndex> &global_column_ids, const MultiFileReaderBindData &bind_data,
-	                             const virtual_column_map_t &virtual_columns);
-	virtual ResultColumnMapping CreateColumnMappingByName(ClientContext &context, MultiFileFileReaderData &reader_data,
-	                                                      const vector<MultiFileReaderColumnDefinition> &global_columns,
-	                                                      const vector<ColumnIndex> &global_column_ids,
-	                                                      const MultiFileReaderBindData &bind_data,
-	                                                      const virtual_column_map_t &virtual_columns);
-
 	//! Used in errors to report which function is using this MultiFileReader
 	string function_name;
